Added a few helper functions to help with Token Maintenance
I made the design decision to export raw time types, allowing
any formating decisions to be made outside of this library.
I.E. it should not be within the scope of duty for this libray
to make things look any specific way (except the string method),
it just obtains a new Client token, and provides you with information
to help with obtaining a new client token when necessary.

- String Format method that prints useful information
  about the Token's expiration status.
- TimeTillExpiration, sends time.Duration until the token
  expires.
- IsExpired, Simple boolean showing the token status.